allows user to set the mouse button used to paste
This helps for people using laptops that may not have a middle mouse
button to use. The value can be set via the Xresources variable
"st.st.paste-mouse-button". It can only be set to "right" currently.

This likely can and should be expanded on to account for mice with
buttons greater than 3.